Adding export function
This simplifies code for methods (they don't need to figure out how to return the correct data type) and allows this logic to be centralised.

Though this method supports multiple types of input (all converted to an array) it isn't yet possible to 'inject' other types in without replacing this entire method, this will therefore need to be reconsidered.

Methods that return a single result are not currently handled and still export to an array and then call the export function. This is fine since the export function passes arrays straight through. This will be fixed soon.

The delete method will also be updated to ensure it returns a result in the expected format (of an array).